+++ extension/src/interfaces.ts
@@ -1,12 +1,18 @@
 export interface IExtension {
-  canRunCli: (cwd: string, isCliGlobal?: true) => Promise<boolean>
+  getCliVersion: (
+    cwd: string,
+    isCliGlobal?: true
+  ) => Promise<string | undefined>
   hasRoots: () => boolean
-  isDvcPythonModule: () => Promise<boolean>
+  isPythonExtensionUsed: () => Promise<boolean>
 
   setupWorkspace: () => void
 
   initialize: () => Promise<void[]>
   resetMembers: () => void
+
   setAvailable: (available: boolean) => void
+  setCliCompatible: (compatible: boolean | undefined) => void
   setRoots: () => Promise<void>
+  unsetPythonBinPath: () => void
 }