+++ webview/src/experiments/components/table/TableHeader.tsx
@@ -81,9 +81,9 @@ export const TableHeader: React.FC<TableHeaderProps> = ({
       },
       {
         id: 'update-header-depth',
-        label: 'Set Max Header Depth',
+        label: 'Set Max Header Height',
         message: {
-          type: MessageFromWebviewType.SET_EXPERIMENTS_HEADER_DEPTH
+          type: MessageFromWebviewType.SET_EXPERIMENTS_HEADER_HEIGHT
         }
       }
     ]